added get users with same direction, get specific userlist and further commentation

+++ TecAlliance.Carpool.API/TecAlliance.Carpool.Business/Models/CarpoolUnitDto.cs
@@ -8,14 +8,14 @@ namespace TecAlliance.Carpool.Business.Models
 {
     public class CarpoolUnitDto
     {
-        public long Id { get; set; }
+        public int Id { get; set; }
         public int PassengerCount { get; set; }
         public string Destination { get; set; }
         public string StartLocation { get; set; }
         public string Departure { get; set; }
-        public List<long> Passengers { get; set; }
+        public List<int> Passengers { get; set; }
 
-        public CarpoolUnitDto(long id, int passengercount, string destination, string startlocation, string departure, List<long> passengers)
+        public CarpoolUnitDto(int id, int passengercount, string destination, string startlocation, string departure, List<int> passengers)
         {
             Id = id;
             PassengerCount = passengercount;
