+++ TecAlliance.Carpool.API/TecAlliance.Carpool.Data/Models/CarpoolUnit.cs
@@ -8,14 +8,14 @@ namespace TecAlliance.Carpool.Data.Models
 {
     public class CarpoolUnit
     {
-        public long Id { get; set; }
+        public int Id { get; set; }
         public int PassengerCount { get; set; }
         public string Destination { get; set; }
         public string StartLocation { get; set; }
         public string Departure { get; set; }
-        public List<long> Passengers { get; set; }
+        public List<int> Passengers { get; set; }
 
-        public CarpoolUnit(long id, int passengercount, string destination, string startlocation, string departure, List<long> passengers)
+        public CarpoolUnit(int id, int passengercount, string destination, string startlocation, string departure, List<int> passengers)
         {
             Id = id;
             PassengerCount = passengercount;
